+++ changelog.md
@@ -3,7 +3,6 @@
  - Base errors support added (i.e. errors flow through all kind of transformations/combinations)
  - Emitter now has `.error` method for emitting errors
  - Emitter object also now has `.error` method
- - Properties now might have a current error along with current value
  - Properties now may have a current error (as well as current value)
  - New method `Kefir.constantError`
  - New method `.mapErrors`
@@ -11,6 +10,9 @@
  - New method `.endOnError`
  - New method `.errorsToValues`
  - New method `.valuesToErrors`
+ - New methos `.skipErrors`
+ - New methos `.skipValues`
+
 
 ## 0.4.2